Rename signal related symbols
Changes:
* connect_builder() -> builder()
* GodotDeref -> UniformObjectDeref
